rename app file
