+++ src/renderer/src/index.tsx
@@ -2,7 +2,7 @@ import { createRoot } from 'react-dom/client'
 
 import './index.css'
 
-import { App } from './App'
+import { App } from './app'
 
 const root = createRoot(document.getElementById('app') as HTMLElement)
